Open control pipe in blocking mode.
Up to now, Pi-FM-RDS was opening the FIFO in non-blocking mode, but with the latest Raspbian, the behavior was: the FIFO had to be opened by the writer before Pi-FM-RDS was started, and not closed after, as pointed out by wmcgrathled. I'm not sure if it was always like this, or if the semantics of opening a FIFO in non-blocking mode changed.

Anyway, it's probably less confusing to open the FIFO in blocking mode, so at Pi-FM-RDS startup, the user is prompted to start the writer process if they haven't done so yet. This commit is changing this.

Fixes #154